Selects activity from backend.
Selects an activity from the back end database based of the logged in user's ID (7 currently, for testing, will be changed later), then shows it on the correct day.
Multiple activities can be shown on one day, when the day is selected, a modal is shown with the title, description, location, url and time shown for each activity.